+++ src/core/errors/EntityIdMismatchError.ts
@@ -5,7 +5,14 @@ import { BusinessRuleViolationError } from './BusinessRuleViolationError';
  * @copyright Piggly Lab 2025
  */
 export class EntityIdMismatchError extends BusinessRuleViolationError {
-	constructor() {
+	/**
+	 * Create a new instance of the error.
+	 * Code: 3612039619
+	 *
+	 * @since 4.0.0
+	 * @author Caique Araujo <caique@piggly.com.br>
+	 */
+	public constructor() {
 		super(
 			'EntityIdMismatchError',
 			'Entity ID mismatch.',