+++ src/core/services/LoggerService.ts
@@ -1,5 +1,15 @@
 /* eslint-disable no-console */
-import type { LoggerServiceSettings } from './types';
+import debug from 'debug';
+
+import {
+	LoggerServiceSettingsSchema,
+	LoggerServiceSettings,
+	LoggerServiceEntry,
+	LogLevel,
+} from '@/core/services/schemas';
+import { OnGoingPromisesService } from '@/core/services/OnGoingPromisesService';
+import { FileLogStreamService } from '@/core/services/FileLogStreamService';
+import { displayLog } from '@/utils';
 
 import { ServiceProvider } from '../ServiceProvider';
 
@@ -8,6 +18,28 @@ import { ServiceProvider } from '../ServiceProvider';
  * @copyright Piggly Lab 2024
  */
 export class LoggerService {
+	/**
+	 * The file log stream service.
+	 *
+	 * @type {FileLogStreamService}
+	 * @protected
+	 * @memberof LoggerService
+	 * @since 4.1.0
+	 * @author Caique Araujo <caique@piggly.com.br>
+	 */
+	protected _file?: FileLogStreamService;
+
+	/**
+	 * Ongoing promises.
+	 *
+	 * @type {OnGoingPromisesService}
+	 * @protected
+	 * @memberof LoggerService
+	 * @since 4.1.0
+	 * @author Caique Araujo <caique@piggly.com.br>
+	 */
+	protected _ongoing: OnGoingPromisesService;
+
 	/**
 	 * Settings.
 	 *
@@ -22,19 +54,71 @@ export class LoggerService {
 	/**
 	 * Constructor.
 	 *
+	 * - alwaysOnConsole: Will display logs on console.
+	 * - callbacks: Custom logger functions:
+	 *   - onDebug: Debug logger.
+	 *   - onError: Error logger.
+	 *   - onFatal: Fatal logger.
+	 *   - onInfo: Info logger.
+	 *   - onWarn: Warn logger.
+	 * - ignoreUnset: When true, will not throw an error if a callback is not set.
+	 * - ignoreLevels: The logger service will ignore any log level set here.
+	 * - onError: Custom error handler.
+	 * - onFlush: Custom flush handler. Should be used to flush logs.
+	 * - file: How to handle file logging
+	 *   - abspath: The path to the file to log to. If not set, the logger will not log to a file.
+	 *   - killOnLimit: If true, will kill the process if the limit is reached.
+	 *   - levels: The levels to log to.
+	 *   - streamLimit: The limit of pending messages for each stream.
+	 * - promises: How to handle promises
+	 *   - killOnLimit: If true, will kill the process if the limit is reached.
+	 *   - limit: The limit of ongoing promises.
+	 *   - track: The callbacks to track.
+	 *
 	 * @public
 	 * @constructor
 	 * @memberof LoggerService
+	 * @throws {ZodError} If settings are invalid.
 	 * @since 4.0.0
+	 * @since 4.1.0 Added ignoreLevels, trackOnGoing.
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
-	constructor(settings: Partial<LoggerServiceSettings> = {}) {
-		this._settings = {
-			alwaysOnConsole: settings.alwaysOnConsole ?? false,
-			callbacks: settings.callbacks ?? {},
-			ignoreUnset: settings.ignoreUnset ?? true,
-			onFlush: settings.onFlush,
-		};
+	public constructor(settings: LoggerServiceEntry = {}) {
+		this._settings = LoggerServiceSettingsSchema.parse(settings);
+		this._ongoing = new OnGoingPromisesService(this._settings.promises);
+
+		if (this._settings.file) {
+			this._file = new FileLogStreamService(this._settings.file);
+		}
+	}
+
+	/**
+	 * The cleanup method will:
+	 *
+	 * - Wait for all pending promises triggered by send method.
+	 * - Clear the pool of pending promises.
+	 *
+	 * Useful for:
+	 *
+	 * - Flush pool of pending promises.
+	 * - Ensure all promises are settled.
+	 * - Gracefully shutdown your application.
+	 *
+	 * @returns {Promise<void>}
+	 * @public
+	 * @since 4.1.0
+	 * @memberof LoggerService
+	 * @author Caique Araujo <caique@piggly.com.br>
+	 */
+	public async cleanup(): Promise<void> {
+		debug('logger:cleanup')('cleaning up');
+		await this._ongoing.cleanup();
+
+		if (this._file) {
+			this._file.cleanup();
+		}
+
+		debug('logger:cleanup')('cleaned up');
 	}
 
 	/**
@@ -49,6 +133,11 @@ export class LoggerService {
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
 	public debug(message?: string, ...args: any[]): void {
+		if (this._settings.ignoreLevels.includes('debug')) {
+			return;
+		}
+
+		this.onFileLog('debug', message, ...args);
 		return this.prepare('onDebug', message, ...args);
 	}
 
@@ -64,6 +153,11 @@ export class LoggerService {
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
 	public error(message?: string, ...args: any[]): void {
+		if (this._settings.ignoreLevels.includes('error')) {
+			return;
+		}
+
+		this.onFileLog('error', message, ...args);
 		return this.prepare('onError', message, ...args);
 	}
 
@@ -79,6 +173,11 @@ export class LoggerService {
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
 	public fatal(message?: string, ...args: any[]): void {
+		if (this._settings.ignoreLevels.includes('fatal')) {
+			return;
+		}
+
+		this.onFileLog('fatal', message, ...args);
 		return this.prepare('onFatal', message, ...args);
 	}
 
@@ -96,14 +195,31 @@ export class LoggerService {
 			return;
 		}
 
-		this._settings.onFlush().catch(error => {
-			if (!this._settings.onError) {
-				console.error('LoggerService.UncaughtError', error);
-				return;
-			}
+		if (this._settings.promises.track.includes('onFlush') === false) {
+			this._settings.onFlush().catch(error => {
+				if (!this._settings.onError) {
+					console.error('LOGGER/UNCAUGHT_ERROR', error);
+					debug(`logger:uncaught`)(error);
+					return;
+				}
 
-			this._settings.onError(error);
-		});
+				this._settings.onError(error);
+			});
+
+			return;
+		}
+
+		this._ongoing.register(
+			this._settings.onFlush().catch(error => {
+				if (!this._settings.onError) {
+					console.error('LOGGER/UNCAUGHT_ERROR', error);
+					debug(`logger:uncaught`)(error);
+					return;
+				}
+
+				this._settings.onError(error);
+			}),
+		);
 	}
 
 	/**
@@ -118,7 +234,26 @@ export class LoggerService {
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
 	public info(message?: string, ...args: any[]): void {
+		if (this._settings.ignoreLevels.includes('info')) {
+			return;
+		}
+
+		this.onFileLog('info', message, ...args);
 		return this.prepare('onInfo', message, ...args);
+	}
+
+	/**
+	 * Wait for a given amount of milliseconds.
+	 *
+	 * @param {number} ms
+	 * @returns {Promise<void>}
+	 * @public
+	 * @memberof LoggerService
+	 * @since 4.1.0
+	 * @author Caique Araujo <caique@piggly.com.br>
+	 */
+	public wait(ms: number): Promise<void> {
+		return new Promise(resolve => setTimeout(resolve, ms));
 	}
 
 	/**
@@ -133,19 +268,44 @@ export class LoggerService {
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
 	public warn(message?: string, ...args: any[]): void {
+		if (this._settings.ignoreLevels.includes('warn')) {
+			return;
+		}
+
+		this.onFileLog('warn', message, ...args);
 		return this.prepare('onWarn', message, ...args);
+	}
+
+	/**
+	 * Log to file.
+	 *
+	 * @param {LogLevel} level
+	 * @param {string} message
+	 * @param {any[]} args
+	 * @protected
+	 * @memberof LoggerService
+	 * @since 4.1.0
+	 * @author Caique Araujo <caique@piggly.com.br>
+	 */
+	protected onFileLog(level: LogLevel, message?: string, ...args: any[]): void {
+		if (this._file) {
+			this._file.log(
+				level,
+				displayLog(level, message ?? 'Unknown log', args).message,
+			);
+		}
 	}
 
 	/**
 	 * Prepare the logger.
 	 *
 	 * @param {'onDebug' | 'onError' | 'onFatal' | 'onInfo' | 'onWarn'} callback
-	 * @param {string} message
-	 * @param {Record<string, any>} meta
+	 * @param {any[]} args
 	 * @returns {void}
 	 * @protected
 	 * @memberof LoggerService
 	 * @since 4.0.0
+	 * @since 4.1.0 Added uncaught error handler.
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
 	protected prepare(
@@ -183,18 +343,33 @@ export class LoggerService {
 			}
 		}
 
-		if (!fn) {
+		debug(`logger:${callback}`)(args);
+
+		if (this._settings.promises.track.includes(callback) === false) {
+			fn(...args).catch(error => {
+				if (!this._settings.onError) {
+					console.error('LOGGER/UNCAUGHT_ERROR', error);
+					debug(`logger:uncaught`)(error);
+					return;
+				}
+
+				this._settings.onError(error);
+			});
+
 			return;
 		}
 
-		fn(...args).catch(error => {
-			if (!this._settings.onError) {
-				console.error('LoggerService.UncaughtError', error);
-				return;
-			}
+		this._ongoing.register(
+			fn(...args).catch(error => {
+				if (!this._settings.onError) {
+					console.error('LOGGER/UNCAUGHT_ERROR', error);
+					debug(`logger:uncaught`)(error);
+					return;
+				}
 
-			this._settings.onError(error);
-		});
+				this._settings.onError(error);
+			}),
+		);
 	}
 
 	/**
@@ -247,9 +422,12 @@ export class LoggerService {
 			return ServiceProvider.resolve('LoggerService');
 		}
 
+		// @note it will not log anything
 		return new LoggerService({
 			alwaysOnConsole: false,
+			ignoreLevels: ['debug', 'info', 'warn', 'error', 'fatal'],
 			ignoreUnset: true,
+			promises: { killOnLimit: false, limit: 10000, track: [] },
 		});
 	}
 }