+++ CHANGELOG.md
@@ -168,3 +168,17 @@
 * [Add] `EnvironmentType` type to handle environment types.
 * [Add] `loadConfigIni`, `loadDotEnv` and `loadYaml` (`js-yaml` and `dotenv` are optional dependencies, should be installed when needed) functions to handle configuration files.
 * [Add] `sanitizeRecursively` function to sanitize a param recursively.
+
+## 4.1.0 at `2025-04-14`
+
+* [Update] Overall dependecies updates;
+* [Change] `InvalidSchemaError` now extends `BusinessRuleViolationError`;
+* [Change] Options for `LoggerService` class without breaking changes;
+* [Fix] Overall typing issues;
+* [Fix] `toJSON` method must have optional parameters at `ApplicationError` class;
+* [Add] `debug` dependecy to display debug logs;
+* [Add] `FileLogStreamService` class to handle file logging with `LoggerService`;
+* [Add] `OnGoingPromisesService` class to handle ongoing promises on `LoggerService`;
+* [Add] `evaluateAbspath` and `parseAbspath` utility functions to handle abspath;
+* [Add] `displayLog` utility function to handle log display;
+* [Add] Schemas for services settings.